Move Server to its own module
